pyproject.toml
[tool.poetry]
name = "pyvem"
version = "1.0.0"
description = "Manage pipenv, poetry env, virtualenv and more virtual envs from one place."
authors = ["Jiri Kyjovsky <j1.kyjovsky@gmail.com>"]
maintainers = ["Jiří Kyjovský <j1.kyjovsky@gmail.com>"]
license = "GPLv3"
readme = "README.md"
homepage = "https://github.com/nikromen/pyvem"
repository = "https://github.com/nikromen/pyvem"
exclude = ["test/"]


[tool.poetry.dependencies]
python = "^3.8"
click = ">=8.0.0"
pyyaml = ">=5.0"
pydantic = ">=1.0"
docker = ">=6.0"
podman = ">=4.0"
tqdm = ">=4.0"
pexpect = ">=4.0"


[tool.poetry.dev-dependencies]
pytest = ">=7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
pyvem = "pyvem.cli.main:entry_point"
